refactor: move the NIP-98 builder to console/, so the page and the CLI share one
The console needs it too, and tools/serve-console.mjs serves only console/ — a
module under src/ is not reachable from the page at all. The alternative was two
copies of a security-relevant builder, which is how two copies drift apart. tests/
already import from console/ for the same reason.

Hashing moves from node:crypto to Web Crypto so the one file runs in both places,
which makes the builder async. Base64 goes through the UTF-8 bytes rather than the
raw string, because btoa throws on a multi-byte character — today every tag value
is a URL or hex, until somebody adds one that is not.

It still does not sign. That began as the egress ban's rule and turns out to be the
right shape regardless: the console signs through the operator's own signer, an
extension or a bunker, never a key the page holds; the CLI signs with a key from a
file. Neither can be baked into a shared builder.
