+++ tools/relay-invite.mjs
@@ -32,7 +32,7 @@ import { dirname, resolve } from 'node:path'
 import { homedir } from 'node:os'
 import { nip19 } from 'nostr-tools'
 import { getPublicKey, finalizeEvent } from 'nostr-tools/pure'
-import { nip98Template, nip98Header, expectedUrl } from '../src/nip98.mjs'
+import { nip98Template, nip98Header, expectedUrl } from '../console/nip98.mjs'
 
 const argv = process.argv.slice(2)
 const cmd = argv[0]
@@ -65,9 +65,10 @@ function readSecret(pathArg) {
 
 async function post(path, bodyObj, secret) {
   const url = expectedUrl(process.env.BUZZ_RELAY_URL, path)
-  // src/nip98.mjs deliberately does not sign — the egress ban keeps `finalizeEvent` out of src/,
-  // so the key stays with the caller that holds it. That is here.
-  const { template, body } = nip98Template({ url, method: 'POST', body: JSON.stringify(bodyObj) })
+  // console/nip98.mjs deliberately does not sign — the console signs through the operator's own
+  // signer and this tool signs with a key from a file, so neither can be baked into the builder.
+  // It lives in console/ because the page is its other caller; see the header there.
+  const { template, body } = await nip98Template({ url, method: 'POST', body: JSON.stringify(bodyObj) })
   const header = nip98Header(finalizeEvent(template, secret))
   let res
   try {
